requirements.txt
streamlit
gspread
oauth2client
pandas
pillow
google-api-python-client
openpyxl
